organizations add role

+++ PrimeApps.Console/Controllers/OrganizationController.cs
@@ -223,6 +223,7 @@ namespace PrimeApps.Console.Controllers
                 Label = organization.Label,
                 Icon = organization.Icon,
                 OwnerId = organization.OwnerId,
+                Default = organization.Default,
                 Teams = organization.Teams.Where(x => !x.Deleted).Select(x =>
                     new TeamModel
                     {
@@ -358,6 +359,9 @@ namespace PrimeApps.Console.Controllers
 
             if (organization == null)
                 return NotFound(ApiResponseMessages.ORGANIZATION_NOT_FOUND);
+
+            if (!await _permissionHelper.CheckUserRole(AppUser.Id, organization.Id, OrganizationRole.Administrator))
+                return Forbid(ApiResponseMessages.PERMISSION);
 
             await _organizationRepository.Delete(organization);
